+++ src/main/java/com/sbhandare/pawdopt/Repository/SecurityUserRepository.java
@@ -3,6 +3,6 @@ package com.sbhandare.pawdopt.Repository;
 import com.sbhandare.pawdopt.Model.SecurityUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SecurityUserRepository extends JpaRepository<SecurityUser,Integer> {
+public interface SecurityUserRepository extends JpaRepository<SecurityUser, Long> {
     SecurityUser findByUsername(String username);
 }
