+++ src/main/java/com/sbhandare/pawdopt/Service/PetService.java
@@ -6,9 +6,9 @@ import com.sbhandare.pawdopt.DTO.PetDTO;
 import java.util.List;
 
 public interface PetService {
-    PetDTO getPetById(int id);
+    PetDTO getPetById(long id);
     PageDTO getAllPets(int page);
     List<PetDTO> getPetsByUsername(String username);
-    List<PetDTO> getPetsByOrgId(int orgid);
-    int savePet(PetDTO petDTO, int orgid);
+    List<PetDTO> getPetsByOrgId(long orgid);
+    long savePet(PetDTO petDTO, long orgid);
 }
