+++ src/main/java/com/sbhandare/pawdopt/Util/PawdoptConstantUtil.java
@@ -2,7 +2,7 @@ package com.sbhandare.pawdopt.Util;
 
 public class PawdoptConstantUtil {
     public static final String ADOPTABLE = "adoptable";
-    public static final int NO_SUCCESS = -1;
+    public static final long NO_SUCCESS = -1;
     public static final int DEFAULT_PAGE_RESULTS = 10;
     public static final String ROLE_USER = "ROLE_USER";
     public static final String ROLE_ADMIN = "ROLE_ADMIN";
